iterator test: use sinks everywhere
The original code was writen about 7 years ago. At that time the Go
compiler was not yet smart enough to optimize out too much code with no
observable side-effetcs. The situation has changed so this commit adds
assignments of the K's and V's to static sink variables for all of the
iterator tests.